requirements.txt
fastapi
uvicorn
python-multipart
motor
streamlit-autorefresh
deep-translator
openai-whisper
transformers==4.38.2
torch
pydantic
pydub
python-dotenv
streamlit
requests
yake
resemblyzer
spectralcluster
numpy
